+++ MovieTickets/Controllers/CinemaController.cs
@@ -2,11 +2,14 @@
 using Microsoft.AspNetCore.Mvc;
 
 using MovieTickets.Services.Data.Interfaces;
+using MovieTickets.Web.Infrastructure.Extensions;
 using MovieTickets.Web.ViewModels.Cinema;
+
+using static MovieTickets.Common.NotificationsConstant;
 
 namespace MovieTickets.Web.Controllers
 {
-    [Authorize(Roles = "Admin")]
+    [Authorize]
     public class CinemaController : Controller
     {
         private readonly ICinemaService cinemaService;
@@ -33,7 +36,12 @@ namespace MovieTickets.Web.Controllers
         [HttpGet]
         public IActionResult Create()
         {
-            return View();
+            if (User.Role() == "Admin")
+            {
+                return View();
+            }
+
+            return RedirectToAction("AccessDenied", "Account");
         }
 
         [HttpPost]
@@ -44,7 +52,16 @@ namespace MovieTickets.Web.Controllers
                 return View();
             }
 
-            await cinemaService.AddCinemaAsync(cinemaModel);
+            try
+            {
+                await cinemaService.AddCinemaAsync(cinemaModel);
+            }
+            catch (Exception)
+            {
+                ModelState.AddModelError(string.Empty, "Error on Create a Cinema, Please try again");
+            }
+
+            TempData[SuccessMessage] = "Create a Cinema Complete";
 
             return RedirectToAction("AllCinemas");
         }
@@ -53,28 +70,50 @@ namespace MovieTickets.Web.Controllers
         [AllowAnonymous]
         public async Task<IActionResult> Details(int id)
         {
-            var detailsModel = await cinemaService.GetCinemaByIdAsync(id);
-
-            if (detailsModel == null)
+            try
             {
-                return View("NotFound");
+                var detailsModel = await cinemaService.GetCinemaByIdAsync(id);
+
+                if (detailsModel != null)
+                {
+                    return View(detailsModel);
+                }
+            }
+            catch (Exception)
+            {
+                TempData[ErrorMessage] = "Error on Details Cinema, Please try Again";
+
+                return RedirectToAction("AllCinemas");
             }
 
-            return View(detailsModel);
+            return View("NotFound");
         }
 
         [HttpGet]
         public async Task<IActionResult> Edit(int id)
         {
-
-            var cinemaEdit = await cinemaService.GetCinemaByIdAsync(id);
-
-            if (cinemaEdit == null)
+            if (User.Role() == "Admin")
             {
+                try
+                {
+                    var cinemaEdit = await cinemaService.GetCinemaByIdAsync(id);
+
+                    if (cinemaEdit != null)
+                    {
+                        return View(cinemaEdit);
+                    }
+                }
+                catch (Exception)
+                {
+                    TempData[ErrorMessage] = "Error on Edit Cinema, Please try Again";
+
+                    return RedirectToAction("AllCinemas");
+                }
+
                 return View("NotFound");
             }
 
-            return View(cinemaEdit);
+            return RedirectToAction("AccessDenied", "Account");
         }
 
         [HttpPost]
@@ -84,7 +123,17 @@ namespace MovieTickets.Web.Controllers
             {
                 return View(cinemaModel.Id);
             }
-            await cinemaService.UpdateCinemaAsync(cinemaModel);
+
+            try
+            {
+                await cinemaService.UpdateCinemaAsync(cinemaModel);
+            }
+            catch (Exception)
+            {
+                ModelState.AddModelError(string.Empty, "Error on Create a Cinema, Please try again");
+            }
+
+            TempData[SuccessMessage] = "You Edit Cinema Complete";
 
             return RedirectToAction("AllCinemas");
         }
@@ -92,9 +141,25 @@ namespace MovieTickets.Web.Controllers
         [HttpGet]
         public async Task<IActionResult> Delete(int id)
         {
-            await cinemaService.DeleteCinemaAsync(id);
+            if (User.Role() == "Admin")
+            {
+                try
+                {
+                    await cinemaService.DeleteCinemaAsync(id);
+                }
+                catch (Exception)
+                {
+                    TempData[ErrorMessage] = "Error on Delete Cinema, Please try Again";
 
-            return RedirectToAction("AllCinemas");
+                    return RedirectToAction("AllCinemas");
+                }
+
+                TempData[SuccessMessage] = "You Delete Cinema Complete";
+
+                return RedirectToAction("AllCinemas");
+            }
+
+            return RedirectToAction("AccessDenied", "Account");
         }
     }
 }
